Split out requestSitePost
